app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy>=1.26
sqlalchemy>=2.0
psycopg[binary]>=3.1
pydantic-settings>=2.0
python-multipart
